Bug#19642978: NDB_RESTORE CORES ON BUILT-IN PRIMARY KEY + STAGING BLOB CONVERSIONS ON SPARC
ndb_restore crashes while restoring a table which contains both 
1) a built-in conversion on the primary key and 2) a staging conversion 
on a TEXT column.

During staging, a blob table is created with a primary key column of 
the target type. However, a convert function is not provided to convert 
the primary key values before loading them into the staging blob table. 
This results in corrupted primary key values in the staging blob table. 
Since the test case is a conversion from int to mediumint, the data 
corruption is visible only on a big-endian architecture.

While moving data from the staging table to the target table, the blob 
read fails because it cannot find the primary key in the blob table. 
This results in an ndb_restore core.

Added a fix to check all blob tables to see if there are conversions on 
PKs of their main tables. This check is done after all the main tables 
are processed, so that convert functions and parameters have been set 
for the main tables. The convert function and parameters which are 
used for the primary key in the main table are duplicated in the blob 
table. Added a test case to reproduce the failure on x86.