This is a patch for the bug#19182009 - 5.7 TRIGGERS HANDLE NOT NULL DIFFERENTLY TO OLDER VERSIONS; BREAKS COMPATIBILITY
For a case when a user executes the statement INSERT INTO for the table having
NOT NULL column, the value for this column isn't specified in INSERT statement
and there is BEFORE trigger with action type different from ON INSERT then the
server issues the ERROR 1048 "Column cannot be null” instead of
Warning 1364 "Field doesn't have a default value".

This bug can be illustrated by the following example:

CREATE TABLE t1(a INT, b INT NOT NULL);
CREATE TRIGGER t1_trg_after_del AFTER DELETE ON t1 FOR EACH ROW SET @a := 1;
INSERT INTO t1(a) VALUES (1);
SHOW WARNINGS;

The reason for this bug was that the function command_invokes_insert_triggers()
returned true when it’s called for INSERT statement regardless of the presence
of a trigger with timing BEFORE and action ON INSERT. This led to calling of
the function call_before_insert_triggers() that set temporary NULL null value
for every field in the table being operated. Later on call to the function
check_record() every field in the table was checked on violation of NOT NULL
constraint calling the method Field::check_constraint(). Since the column
declared as NOT NULL was set to temporary NULL value inside the function
call_before_insert_triggers() the method Field::check_constraint() returned
the value true and set error the ERROR 1048 "Column cannot be null” in
the Diagnostic_area.

To fix this bug additional check has been added to guarantee that the function
call_before_insert_triggers() is called only when actually there is a
BEFORE INSERT trigger for the table specified in INSERT statement .

As part of this patch the function command_invokes_insert_triggers() has been
renamed to command_can_invoke_insert_triggers().