Bug #20031243 CREATE TABLE FAILS TO CHECK IF FOREIGN KEY COLUMN NULL/NOT NULL MISMATCH
Problem:

In an alter table command, when we modify the name of a column that is part of
a foreign key constraint, then reloading of the foreign key constraint failed.
This problem happened because the foreign object with old column names are
there in the dictionary cache.

Solution:

To avoid problem, when a column name is changed, then evict the corresponding
foreign key object from the dictionary cache.  It will be loaded subsequently
with foreign keys are loaded.

rb#7369 in review