WL#7083 step 4.6. Simple checks: Wrong event type in running sender thread.
@rpl_binlog_sender.cc
- Introduce check_event_type function, and check that event type does
not contradict gtid_mode or auto_position.

@rpl_binlog_sender.h
- Prototype for new function.
- The check_event_type function calls set_fatal_error. set_fatal_error
currently only copies the pointer to the error message and stores it in
a member variable that is supposed to be used later. But
check_event_type uses a buffer on the stack which will go away later. So
we make check_event_type copy the buffer rather than copy the reference
to the buffer, to prevent using deallocated stack memory.

@errmsg-utf8.txt
- New error messages.