WL#7201, WL#7202, WL#8062: True bottom-up server parser: refactoring of the INSERT, REPLACE, DELETE and UPDATE statements.
This patch implements a few worklogs:

* WL #7201: True bottom-up server parser: refactoring of the INSERT/REPLACE statement
* WL #7202: True bottom-up server parser: refactoring of the DELETE statement
* WL #8062: True bottom-up server parser: refactoring of the UPDATE statement

All 3 worklogs are intended to refactor grammar rules for INSERT, REPLACE,
DELETE and UPDATE statement in a pure bottom-up style to make it (the grammar)
context-independent for the better maintainability and extendability at
the first place.

Worklogs use the framework and the methodology from the WL7199
"True bottom-up server parser: common framework for the refactoring".

Please see WL pages for details.