BUG#19978288 - ASSERT INNOBASE_TABLE->FTS_DOC_ID_INDEX == 0, ALTER
The failing assert looks bogus. The table could be removed away from cache
before this checking, so it should be re-load into cache again. During the
re-load, fts indexes, including fts_doc_id_index will be initialized, which
results in the assert failure.

We should check instead of assert and get the index if necessary.

RB: 7305
Reviewed-by: Marko