BUG#20040791 - ASSERT THR_GET_TRX ERROR_STATE DB_SUCCESS ROW_INS_SEC_INDEX_ENTRY_LOW
For REPLACE statments, once we found a duplicate key error, we will try to
lock the gaps on other unique secondary indexes and set the error_state to
DB_DUPLICATE_KEY. So we assert that the operation should always success.
But it's a bit different for temporary tables. Temporary table doesn't lock
anything, so the error_state won't change. We can simply skip other indexes
if we find a duplicate key on some index of temporary table.

RB: 7367
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>