bug#19930381 WL1697: START SLAVE CRASHES POST RESTART, BLOCKING GTID_MODE (ON/OFF) TESTING
The actual reason of the bug is that MSR removes unrelated to a being reset channel
worker records out of mysql.slave_worker_info.
When previously the table may be totally demolished by RESET SLAVE, now
MSR has to remove affected workers delicately.

That is done with refactoring basic methods of the slave info storage to
chose the key prefix scan instead of former data scanning.

This patch does not address few more lingering issues that are spotted
in MSR+MTS recovery.
To brief, the channel workers are not initialized properly when the worker info
table is read at mts_recovery_groups() time.
Therefore these workers can not take part in recovery.