BUG#16886196 - ALTER TABLE FAILS TO CONVERT TO PREFIX INDEX IN                      ALTER_COLUMN_EQUAL_PACK_LENGTH
Analysis
--------

An assert is observed in a debug build while accessing a
column on which an index is defined after the length of
the column is modified through ALTER TABLE operation.
This issue is observed only during the boundary condition
where:

a) The column length is 767 and has an index defined on it.
b) The column length is extended to a value greater than
   767.

INNODB limits the length of the index to 767 bytes (without
innodb_large_prefix mode enabled). When the ALTER TABLE
operation modifies the column length by extending it to
a value greater than 767, the index is not converted to a
prefix index. This is because the key part length for
the old and the new version of the index remains the
same(767).

The function 'has_index_def_changed()' checks for the key
part length in order to determine whether an index has to
be recreated. Since there is no change in the length, it
assumes there is no change in the index and fails to
convert the index into a prefix index by recreating the
index.

Fix
---

The 'has_index_def_changed()' function now checks, if the
key as has been converted from a non-prefix index to a
prefix index by using the key part and key flags of the
older and newer version of the index respectively.