Bug #19861997  RENAME SIMPLIFIED_BINLOG_GTID_RECOVERY AND EXECUTED_GTIDS_COMPRESSION_PERIOD
In 5.6.21, we introduced the global variable
'simplified_binlog_gtid_recovery', which breaks
the existing naming convention of binlog-related
variables, all starting with "binlog_".

To fix the problem, create a new alias
'binlog_gtid_simple_recovery', and produce a
deprecation warnings on setting and reading the
variable 'simplified_binlog_gtid_recovery'.