Bug#20073459 GIS PERFORMANCE ISSUE: TOO MANY REALLOCS
Geometry data in practical use cases often are at least hundreds of bytes, so
we should use more aggressive allocation increments i.e. this version

         int String::reserve(size_t space_needed, size_t grow_by)

to reserve memory space for GIS data.
This function is dedicated for GIS data, it allocates memory by bigger
chunks, should be called by GIS code. Now when calling this function, we grow
by 512 bytes normally, and only grow by smaller steps if we know for sure we
are reaching the end of the geometry data.

We should not use the int String::reserve(size_t space_needed) version as some
of the GIS code does currently, otherwise we would be reallocating the geometry
memory buffer multiple times incrementally for a single geometry when
generating/producing the geometry data piece by piece, which is a lot of
overhead.

Since this is a performance enhancement, no test cases are added. Existing
GIS tests can verify the correctness of this change.