Bug #20041925: MAKE AUTH_SOCKET MORE FLEXIBLE
Before this fix the unix socket auth plugin returned true only when
the OS socket user id matches the mysql user name.
The authentication string was ignored.

This will still work, but in addition to this (if the comparison fails)
the socket user id is compared to the authentication_string too.

So effectively this plugin will now return a positive if the socket's
user name matches ether the mysql account user name or the mysql account
authentication string now.

Made the plugin to fill in the @@external_user variable.

Fixed a bug in how @@external_user's variable value was formed
(it was a copy of @@proxy_user instead of the value returned by the plugin).

Tests updated to reflect the proper @@external_user value.