BUG#19665003: RPL.RPL_MULTI_SOURCE_BASIC UNSTABLE ON PB2
The test case is a multi-source test case, where server 2 replicates
from server 1 and 3.

There was a race, causing the slave to concurrently execute two CREATE
TABLEs with the same table definition coming from server 1 and 3. One
of the CREATE TABLE statements had an IF EXISTS clause, which would
make the test pass sometimes. However, if that statement was replayed
first on the slave, then next CREATE TABLE would trigger an error
(table already exists).

We fix this by making sure that the CREATE TABLE is always executed
first. We allow the second one to execute as well and be put in the
binary log, since it will not hurt correctness.