Bug#18715165 : NDB_INDEX_STAT FAILS WHEN UNIQUE KEY EXISTS; SEGMENTATION_FAULT
During the process, an array of index name is created. There exists a check that if
the index is an ordered type then only it is been added in the array. However, if
it not an order index the corresponding value of array is skipped resulting, a
garbage value in the array. Accessing these garbage values later in the process,
results in segmentation fault.

This patch make sure that all the names present in the array are corresponding
to an order indexes.

other changes:
  - Added a unit test for it.
  - Added a inc file 
  - update the result file