Bug#20083349: RPL.RPL_PERFSCHEMA_APPLIER_STATUS_BY_COORDINATOR FAILING               IN TEST ASSERTION
The objective of the test was to test that there are no coordinators
on the master. Number of coordinators should be zero. But the assertion
fails sometimes with a value 1.

This test failed sporadically on PB2 but it always failed if it is executed
immmediately after rpl.rpl_dual_pos_advance.test; In the latter
test the configuration is 1->2->1. This means the master is  also
a slave which had it's own SQL thread(coordinator). Because
of this the assertion failed correctly.

The fix was to properly cleanup the rpl_dual_pos_advance.test by
using force_restart.inc