Fixed crashing AutoTest testBitfield.
The test client created a test table with a random number of
columns, where all except the two first (Unsigned) columns
are Bitfield columns. The number of columns were >= 1.

However, there is a requirement in the Hugo framework
that all tables under its control should have a least 
an Unsigned PK column, and *another* unsigned column containing
the update count of the rows.

This fix changes the random table creation such that 
it contains at least these two columns.

It also adds 'requires' to the HugoCalculator c'tor that this
rule is followed, and propper in init of m_idCol and m_updatesCol
which were missing.