Experimental fix for several AutoTest which crash with an unreadable stack dump very early in their lifecycle, like:
- testIndex -l 2 -n SR1_O T6 T13
- testBasicAsynch -n PkDeleteAsynch
- testBasicAsynch -n PkUpdateAsynch
- testScan -n ScanRead40 -l 100 T1 D1 D2

We have not succeeded in reproducing these failures
outside the ATRT framework. However, the crashes looks
very much like stack overruns due to an insufficient
thread stack size.

This fix sets the default thread stack size in the NDBT_Test
framework to 512K, instead of using the common default stack size
in NdbThread_Create which is either 64/128K.

This will *not* affect the thread stack size for threads running
inside the datanodes!

Pushes this to 7.4 only. Will later backport 7.1-> if
it seems to fix the crashes.