Bug #20145425 	INNODB: FIL_SPACE_GET_FLAGS() RETURNS INCORRECT VALUE WHEN PAGE-SIZE IS NOT 16K
The space flags of the system tablesapce is not zero when
innodb-page-size is something other than 16k.  So do not
always return zero for the system tablesapce.  This bug
is not currently seen because this function is either used
on file-per-table tablespaces, or the return value is not
used to get the page size.  But future worklogs 6205 and 7053
have code that expects the correct space flags for the system
tablespace.

Reviewed-by: Marko Mäkelä <marko.makela@oracle.com>
Approved via IM