WL#6819: GCS Replication: Overview - Bug 11
This patch aims to solve Bug 11 - When all the cluster nodes shutdown
bring the cluster up will eventually lead to recovery errors because the
view id will return to 0.

The solution is to add randomness, like a timestamp, at cluster start.
When a node is alone in a cluster, it would create a random id that
would identify that cluster until it goes down (name proposals are
accepted). That id would then be concatenated to a monotonic view
identifier. Each node that enters in the Cluster would receive that
"cluster id".