BUG#20160942: SESSION_TRACK_SYSTEM_VARIABLES_BASIC MAY FAIL SPORADICALLY
The test selects from information_schema.global_variables
to find out which have the prefix "session_track" in the
name. However, the SELECT does not explicitly order the
result by variable name, which may render the test flaky.

We fix this by deploying an ORDER BY clause to the select
operation.