Bug #19784641 WL1697: POST RESTART 'RESET SLAVE ALL' CRASHES
Crash was caused by attempt to reset a default channel whose mi member
was not initted.

RESET-SLAVE handler is refined to expect NULL value and not attempt to
go any further in such case.  Unlike the default channel others can
not have mi as NULL, asserts guarding that in reset_slave() are
reinforced with resetting the mi member value in case initialization
turns badly.