Fix for various AutoTest 'testIndex ...' crashing due to excessive memory usage.
The verifyIndex() utility used by several of the testIndex testcases
opened a NdbIndexScanOperation for every rows fetched from
a NdbScanOperation.

That NdbIndexScanOperation was not 'closed' at the end of its
lifecycle, and was thus not cleaned up until the transaction, and the
enclosing NdbScanOperation was closed/cleaned.

Thus a waste amount of scan buffer memory us occupied, and tests
are failing due to this.

NOTE: This patch alone is not sufficient to fix the failing AutoTest.
The bug 20166585 has been filed in addition, regarding that a ::close()
should release the scan buffer memory.