Update 'v3' fix for bug#20166585:
  EXCESSIVE MEMORY USAGE: NDBSCANOPERATION::CLOSE() DOES NOT RELEASE BUFFER MEM.

The 'scan buffer', used to store rows fetched from a scan batch,
can easily become rather large (Multi MB). It is thus important
that we dont keep this structure longer than necessary.

The ::close() of a NdbScanOperation terminates navigation of
the scan result set. Thus, this is a natural place to release 
this scan buffer.

This patch moves the scan buffer release point from 
NdbScanOperation::release to ::close(). 

NOTE: ::release() terminates the lifecycle of the 
ScanOperation itself. ::close() terminates the lifecycle
of the result set - Thus, ::close() is the natural place to
release such buffer memory.