Follow up to fix for bug#bug#20166585:
  EXCESSIVE MEMORY USAGE: NDBSCANOPERATION::CLOSE() DOES NOT RELEASE BUFFER MEM.

Frazer had some concerns that the delete of m_scan_buffer is
done by ::close_impl while we hold a poll_guard on the client.

This lock was not required for doing the delete, so the 
delete operation is moved from ::close_impl()
into ::close() to a place immediately after the
return from close_impl(). (and the release of poll_guard.)