Bug #20108866: --LOG_TIMESTAMPS=UTC NEEDS --LOG-BIN
The problem was that the --log_timestamp startup option introduced by WL#6661 by mistake
required --log_bin in order to be accepted.

This fix removed the copy/paste of OPT_BINLOG_FORMAT as a required argument from the
command line parsing. Before this fix you would get the folowing warning printed, if you
tried to set --log_timestamps=UTC and log-bin was not used:

"You need to use --log-bin to make --binlog-format work."