Bug #19309652 MAGICALLY REAPPEARING GRANT OPTION FOR PROXY
Description:
-----------
The GRANT PROXY syntax is overloaded to accept a WITH GRANT OPTION
modifier,
which allows the specified user to issue GRANT PROXY privileges for the
proxied user account.  When GRANT PROXY is issued against an existing
proxy/proxied user combination, the operation appears to succeed, but
apparently only affects the in-memory privileges.  Consequentially,
while the
WITH GRANT OPTION appears to have been honored, it is reverted the next
time
FLUSH PRIVILEGES is issued, or the server is restarted

Fix:
------------
Option with_grant was not saved to the disk  when the proxy grant was
updated. This is why it was as if it had not changed at all after
flushing the privilages. The fix makes sure that with_grant gets stored
on the disc.

Testing:
------------
Added new test: i_main.grant-bug19309652