Bug#20198726 ASSERT DICT_INDEX_IS_CLUST(INDEX) || !DICT_INDEX_IS_ONLINE_DDL(INDEX)
A debug assertion failed shortly after ADD INDEX,
because the to-be-added index carried the
index->online_status = ONLINE_INDEX_ABORTED_DROPPED.

This type of indexes should only exist in the data dictionary cache, not in
the persistent data dictionary or on disk. So, it seems that we should not
have any record locks pointing to that index. Or at least there should not be
any new locks. Maybe the locks were enqueued when the index existed as
ONLINE_INDEX_COMPLETE inside InnoDB, but the ALTER TABLE ... ADD INDEX was
rolled back for example due to a MDL upgrade timeout.

The failure occured when we are executing btr_root_raise_and_insert()
on the clustered index. Its root page (0:1093) was split into the two
child pages (0:769) and (0:770). (I checked the root page in the core
dump, it contains two node pointers with these page numbers.)

The root page of the secondary index was 0:769. This is what trips the
debug assertion.

This is a rather transient failure, because the record lock objects
would be removed at transaction commit.

At first sight, the bug seems to be that when we discarded the
secondary index, we should also have removed all record locks on it.

A much simpler fix is to relax the debug assertion, allowing the
ONLINE_INDEX_ABORTED_DROPPED state when records are being moved around
by lock_move_rec_list operations.

Because the online_status of an uncommitted index can be changed from
ONLINE_INDEX_COMPLETE to ONLINE_INDEX_ABORTED or
ONLINE_INDEX_ABORTED_DROPPED virtually at any time, relaxing the
assertion seems to make sense also in other code paths than the one
where the assertion was triggered. We do have assertions in upper
layers that prevent the use of an aborted index. This lock-addition is
just a maintenance operation that is invoked when records are being
moved around.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>