BUG#20029625 - HANDLE_FATAL_SIGNAL (SIG=11) IN DICT_MEM_TABLE_COL_RENAME_LOW
When we try to rename a column which is involved in some foreign key
constraint, and add a new foreign key constraint for it, we have to deal with
the case that the old foreign index could has been dropped. So while renaming
the columns in foreign keys, we should check the indexes and might find the
equivalent indexes for the foreign keys.

rb: 7544
Reviewed by: Sunny