Bug#20031425 NDBINFO NAMESPACE FREE CHECK FAILURE
The SQL script which creates and upgrades the ndbinfo database, tables and
views are in scripts/mysql_system_tables.sql. The script
is run as part of mysql_install_db and mysql_upgrade.
      
The script detects different preconditions to deterine if the ndbinfo objects
should be created or not. For example the mysqld must have been compiled with
ndbinfo handler, the version must be >= 7.1 and also the "ndbinfo namespace"
must be free.
      
The above mention "ndbinfo namespace" check is now failing. This causes the
ndbinfo objects not to be upgrade properly. The "ndbinfo namespace" check is
intended to avoid dropping any user data tables potentially created in the
ndbinfo database.
        
The problem started when two MyISAM "lookup" tables named
ndb$dblqh_tcconnect_state and ndb$dbtc_apiconnect_state was added to the
ndbinfo database.
      
- fix by removing the "ndbinfo namespace" check since only the SQL commands uses
  CREATE IF NOT EXISTS and only drops tables and views by name.
      
- add test case which shows that the upgrade actually does something.