Incremental addendum patch for bug#20112981
   TransporterFacade::reset_send_buffer might reset a send_buffer in use by 'send'

Instead of introducing the method is_locked_send() used only in ::reset_send_buffer(),
the patch is modified to temporary try_lock and unlock the 'send' lock.

Improves the readability of the locking mechanism pr. reviewer comment.
Performance should not be an issue in the reset eiter, called once at startup
and disconnect/reconnect.