Bug#20220909 - OPTIMIZE IBUF MERGE WHEN READING A PAGE FROM DISK
Problem:
--------
For every page we read, we check if there are pending ibuf entries.
We acquire RW_X_LATCH on the IBUF page, check if there is ibuf
entry for the page. Change buffer is applicable only on for B-tree
secondary leaf pages in non-temporary tablespace. So for other page
types, we unnecessarily acquire & release latch on IBUF page (page 1).

Fix:
---
We can skip the ibuf merge early without acquiring & releasing the latch
on IBUF page.

We will use the page_type, page_level & the tablespace type from the
page which is read from disk to skip the ibuf merge.

Reviewed-by:  Marko Makela <marko.makela@oracle.com>