Bug #19845913 MTS: WITH MSR, SLAVE CRASHES ON SHUTDOWN COMMAND AT RPL_MTS_SUBMODE.H:35
The crashes were caused by an attempt to access already deleted
info by Slave SQL thread or Worker threads.
WL6314 design time deallocation of rli->current_submode turned to be wrong.
It actually was against existing pattern of how to do that properly.
In more details the shutdown thread may break waiting for slave threads
explicit exit through timeout option.
And when it choses that it carries on to destoy the above object
which can be still in use. That causes crashing.

Proper deletion of dyn objects related to MTS, RLI::current_submode
incl, must be done inside or near slave_stop_workers().
And that's what this patch has to correct.

I could not create an mtr testcase because testing with a mysqltest
client connection that does return ack before the server gets down
is something that mtr does not like:
Attempt to execute

  --send STOP SLAVE

  let $restart_file= $MYSQLTEST_VARDIR/tmp/mysqld.1.expect;
  --exec echo "wait" > $restart_file
  --shutdown_server 10
  ...

results in
  mysqltest failed but provided no output

Therefor the patch is tested manually.