BUG#20128071 - WL8202: SERVER CRASHED ON SETTING GCS_REPLICATION_PLUGIN_COMPONENTS_STOP_TIMEOUT
Problem : The setting of the value of the
gcs_replication_plugin_components_stop_timeout was causing the server
crash since the recovery_module was not initialized without the start
gcs_replication.

Solution : The setting of the timeout value is now guarded by the check
of whether the recovery_module is initialized or not.