BUG#19704710: TWO DIFFERENT REPLICATION CHANNELS WITH SAME (HOST, PORT) ARE POSSIBLE
MSR was implemented taking into consideration that there shall be a
1-1 mapping between a replication channel and {host,port} pair. This
was implemented by string comparing the (host,pair) of a particular
channel from the CHANGE MASTER with the already existing (host,port) pair.
The implementation overlooked the fact that host could be hostnames or
ip addresses.
Having a 1-1 correspondence between a channel and (host,port) prevents
several use cases for users. So, it is decided to remove the restriction.

The fix for the bug was not to have any mapping between a channel and
 (host,port) pair. This results in undoing of the following changes
in the code:
1. Remove the check that compares the (host,port) taken from CHANGE MASTER
   with the existing (host,port) pairs
2. The tests which were added to check 1.

After this fix,
It is possible to create two channels to the same source.