WL#6834: GCS Replication- Extract PKE from RBR events for certification purposes.
WL#6834: GCS Replication- Extract PKE from RBR events for certification
purposes

This task designs and implements a mechanism to extract a primary key
equivalent (PKE) from the existing information shipped in RBR. This
information is then used to check for conflicts between transactions. We
could use the after image (AI) and/or (BI) values, but that would
probably not be optimal w.r.t. memory usage. Instead, we could perhaps
calculate hashes and handle those.

This patch also introduces a new server option to decide when to have
the write set generated and when should not.