Bug#20059644: WL6737:CRASH AT OPERATOR<< IN DICT/DICT0MEM.CC
In latest wl6737, the bug isn't crashed any more. It reports
ERROR 1034 (HY000): Incorrect key file for table '/tmp/#sql_4266_0'; try to
repair it

The error is because MySQL doesn't remove the duplicated order-by(group-by) list
completely. In current trunk, if the item->type() of order-by(group-by) is not
Item_field, optimizer skips to remove duplicated items. This doesn't make sense.
When internal tmp table is used, Innodb doesn't support to create a key on
duplicated columns.

So this bug fix removes the duplicated order-by(group-by) items completely
during query optimization.

BTW, because there is an error when compare two weight_string functions to be
equal or not, this patch fixes this problem too.