Bug #19975322 "ROW SIZE TOO LARGE" ERROR WHEN INSERT NOT WITH UPDATE COMMAND
PROBLEM

In Innodb we enforce a limit that the BLOB data inserted in one operation
cannot be greater than 10% of total redo log size. This is done to ensure that
checkpoint lsn is not overwritten. This is the reason we are getting the error
during inserts when we try to insert in all blob fields,but during update we
we were updating one blob field at a time and not hitting this limit. The user
though doesn't get a proper error message which leads to confusion.

FIX

We have improved the error message to give proper information.

[Approved by Jimmy rb#7569]